ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.0
langchain==0.3.0
langchain-google-genai==2.0.0
pydantic==2.9.0
python-dotenv==1.0.1
# EHR Data Extraction dependencies
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.2
python-multipart==0.0.9
